+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 - Added `-b/--template-branch` parameter to `seedfarmer init project` and `seedfarmer init module` so it permits to use multiple branches from a single repository.
+- Added support for downloading modules from archives using HTTPS
 - Added generic module deployment roles for modules without modulestack
 
 ### Changes